backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-Migrate==4.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
PyJWT==2.6.0
passlib==1.7.4
psycopg2-binary==2.9.5
pytest==7.2.0
pytest-cov==4.0.0
python-decouple==3.6
requests==2.28.1
SQLAlchemy~=2.0.9
decouple~=0.0.7